pyproject.toml
[project]
name = "ai-live-commerce"
version = "0.3.1"
description = "AI Live Commerce Platform — Automated live streaming commerce with AI avatar"
readme = "README.md"
requires-python = ">=3.10"
license = { text = "MIT" }

dependencies = [
  "litellm>=1.40.0",
    # Core
    "fastapi>=0.110.0",
    "uvicorn[standard]>=0.27.0",
    "pydantic>=2.6.0",
    "pydantic-settings>=2.1.0",
    "pyyaml>=6.0",
    "python-dotenv>=1.0.0",

    # Database
    "sqlalchemy>=2.0.25",
    "aiosqlite>=0.19.0",

    # Logging
    "structlog>=24.1.0",

    # LLM Providers
    "anthropic>=0.18.0",
    "google-generativeai>=0.4.0",
    "openai>=1.12.0",
    "groq>=0.5.0",

    # Chat Monitoring
    # TikTokLive needs betterproto which only works on Python <3.13
    # On Python 3.13+ this line is SKIPPED by the resolver
    "TikTokLive>=6.0.0; python_version < '3.13'",
    "websockets>=12.0",

    # Voice / TTS
    "edge-tts>=6.1.0",

    # Media Processing
    "ffmpeg-python>=0.2.0",
    "numpy>=1.26.0",
    "Pillow>=10.2.0",

    # HTTP
    "httpx>=0.27.0",
]

[project.optional-dependencies]
gpu = [
    "torch>=2.4.0",
]
livetalking = [
    # LiveTalking dependencies
    "torch>=2.4.0",
    "torchvision>=0.19.0",
    "torchaudio>=2.4.0",
    "opencv-python>=4.9.0",
    "scikit-image>=0.22.0",
    "scipy>=1.12.0",
    "librosa>=0.10.0",
    "soundfile>=0.12.0",
    "resampy>=0.4.0",
    # WebRTC for real-time streaming
    "aiortc>=1.6.0",
    "av>=11.0.0",
    # Additional media processing
    "imageio>=2.33.0",
    "imageio-ffmpeg>=0.4.9",
]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "pytest-cov>=4.1.0",
    "hypothesis>=6.98.0",
    "ruff>=0.3.0",
    "mypy>=1.8.0",
    "pre-commit>=3.6.0",
]

[project.scripts]
serve = "src.main:main"

# ── UV Configuration ──────────────────────────────────
# UV creates .venv INSIDE this project directory (portable!)
# Does NOT touch conda or system Python
[tool.uv]
# .venv is created in: videoliveai/.venv/
# All packages installed there, not in conda or global site-packages
managed = true

[tool.ruff]
target-version = "py310"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP"]

[tool.mypy]
python_version = "3.10"
strict = true
warn_return_any = true
warn_unused_configs = true

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
markers = [
    "integration: requires GPU server (deselect with '-m not integration')",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

